requirements.txt
anthropic
python-docx
python-dotenv
streamlit
requests
beautifulsoup4
lxml
